Add files via upload
This project is analyzing data for an online gaming store Ice, which sells video games globally. The aim is to identify patterns that determine the success of a game or not. Data will be available from open sources including users reviews, expert reviews, genres, platforms and historical data on game sales.