Add files via upload
An analysis will be conducted for the company Zuber. Zuber is a new ride sharing company in Chicago. The aim is to support Zuber in having a successful launch in Chicago. Data is provided from competitors. An examination of passenger preferences and how external factors impact rides will be conducted.